Move tests to npv.py
