--- esculator/tests/base.py
@@ -1,15 +1,0 @@
-from types import FunctionType
-
-
-def snitch(func):
-    """
-        This method is used to add test function to TestCase classes.
-        snitch method gets test function and returns a copy of this function
-        with 'test_' prefix at the beginning (to identify this function as
-        an executable test).
-        It provides a way to implement a storage (python module that
-        contains non-executable test functions) for tests and to include
-        different set of functions into different test cases.
-    """
-    return FunctionType(func.func_code, func.func_globals,
-                        'test_' + func.func_name, closure=func.func_closure)
